Address PR feedback - BasicStatesPlanVisitor and AdvancedStatsPlanVisitor
